requirements.txt
asgi-redis==1.3.0
asgiref==1.1.1
attrs==16.3.0
autobahn==17.5.1
Automat==0.5.0
channels==1.1.3
constantly==15.1.0
coverage==4.3.4
cycler==0.10.0
daphne==1.2.0
Django==1.10
django-mptt==0.8.7
incremental==16.10.1
matplotlib==2.0.0
msgpack-python==0.4.8
numpy==1.12.1
olefile==0.44
pandas==0.19.2
Pillow==4.0.0
pyparsing==2.2.0
pypiwin32==220; sys.platform == 'win32'
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
Twisted==17.1.0
txaio==2.7.1
zope.interface==4.4.0
